Added folder of contents for clarification
The contents in this folder are the same as in count-the-days.tar.gz , but they coudn't be viewed from github, so I decided to upload this folder too. The reason there is a compressed file is because when it's extracted there is immediately a hidden folder created (hidden files cannot be uploaded to github) that can easily be accessed and moved to the home directory.